remove redundant fragment
